+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - Add 'vdiffr' test cases
 
+## Bug fixes
+
+- Fix bug where a blank label causes an error
+
 # ggfittext 0.8.0
 
 ## Major changes